add `/web` and `/node` subpaths
