+++ test/app.test.ts
@@ -1,6 +1,6 @@
 import { Readable, Transform } from "node:stream";
 import { describe, it, expect } from "vitest";
-import { fromNodeHandler } from "../src/node";
+import { fromNodeHandler } from "../src/adapters/node";
 import { eventHandler, createError, setResponseHeader } from "../src";
 import { setupTest } from "./_utils";
